Avoid overvisiting some BazelTargets
If a BuildTarget has already been visisted and it's a library or binary
there is no point of revisiting it.

For others like file or protobuf there is still a need as we need to
deal with includes and deps that need to be added to the current Bazel
target. That's why we continue revisiting.