Prune transitive phony targets
Let's say we have A -> B -> C -> D and B and C are phony builds.
We want to prune B because "it's empty", under *some* circumptances it
would be desirable to prune C so that A depends on D directly but on
some others there is value to keep C so that all the inputs (D) are kept
separated from other deps that A might have.